🛡️ Sentinel: [CRITICAL] Fix API Key Exposure by centralizing environment validation and output redaction
Creates a central `scripts/fleet/env.ts` file that enforces the existence of `JULES_API_KEY` and `GITHUB_TOKEN` across all fleet orchestration scripts at startup. Additionally, it intercepts global `console` output methods (`log`, `error`, `warn`) to safely scrub these sensitive tokens before writing to `process.stdout`/`stderr`. This prevents keys from leaking in execution logs or when external libraries log exceptions containing the environment.

* Creates `scripts/fleet/env.ts` for centralized credential logic
* Updates all entrypoints (`fleet-plan.ts`, `fleet-dispatch.ts`, `fleet-merge.ts`, `fleet-analyze.ts`) to import `env.ts` first
* Cleans up the redundant `github/logging.ts` file
