+++ scripts/fleet/fleet-merge.ts
@@ -15,6 +15,7 @@
 import path from "node:path";
 import { findUpSync } from "find-up";
 import type { IssueAnalysis, Task } from "./types.js";
+import { JULES_API_KEY, GITHUB_TOKEN } from "./env.js";
 import { getGitRepoInfo, getCurrentBranch } from "./github/git.js";
 import { jules } from "@google/jules-sdk";
 
@@ -22,23 +23,11 @@ const repoInfo = await getGitRepoInfo();
 const OWNER = repoInfo.owner;
 const REPO = repoInfo.repo;
 const BASE_BRANCH = process.env.FLEET_BASE_BRANCH ?? "main";
-const GITHUB_TOKEN = process.env.GITHUB_TOKEN;
-const JULES_API_KEY = process.env.JULES_API_KEY;
 
 // Re-dispatch configuration
 const MAX_RETRIES = Number(process.env.FLEET_MAX_RETRIES ?? 2);
 const PR_POLL_INTERVAL_MS = 30_000;
 const PR_POLL_TIMEOUT_MS = 15 * 60 * 1000;
-
-if (!GITHUB_TOKEN) {
-  console.error("❌ GITHUB_TOKEN environment variable is required.");
-  process.exit(1);
-}
-
-if (!JULES_API_KEY) {
-  console.error("❌ JULES_API_KEY environment variable is required.");
-  process.exit(1);
-}
 
 const headers = {
   Authorization: `Bearer ${GITHUB_TOKEN}`,
